Rename image picker widget to image picker page.
