+++ lib/pages/home/dream_tab.dart
@@ -5,9 +5,9 @@ import 'package:stable_horde_flutter/blocs/conversions_bloc.dart';
 import 'package:stable_horde_flutter/blocs/shared_prefs_bloc.dart';
 import 'package:stable_horde_flutter/blocs/tasks_bloc.dart';
 import 'package:stable_horde_flutter/pages/home_page.dart';
-import 'package:stable_horde_flutter/pages/image_picker_page.dart';
 import 'package:stable_horde_flutter/pages/prompt_edit_page.dart';
 import 'package:stable_horde_flutter/pages/seed_edit_page.dart';
+import 'package:stable_horde_flutter/widgets/image_picker_widget.dart';
 import 'package:stable_horde_flutter/widgets/model_button.dart';
 import 'package:stable_horde_flutter/widgets/section_frame.dart';
 
@@ -30,7 +30,7 @@ class _DreamTabState extends State<DreamTab> {
             const SizedBox(height: 16),
             _promptWidget(),
             const SizedBox(height: 32),
-            const ImagePickerPage(),
+            const ImagePickerWidget(),
             const SizedBox(height: 16),
             ExpandablePanel(
               header: const Text(
